requirements.txt
Flask==1.1.2
jinja2<3.1.0
gunicorn==20.1.0
itsdangerous==2.0.1
werkzeug==2.0.2
opencv-python-headless==4.5.5.64
numpy==1.22.3
pillow==8.4.0
flask-cors==3.0.10
imagekitio==2.2.8
python-dotenv==0.19.2
scikit-image==0.18.1